- Moved methods from AbstractCarrierInterface to CarrierInterface - Deprecated AbstractCarrierInterface - Removed references to AbstractCarrierInterface in fthe CarrierFactoryInterface, changed them to CarrierInterface. - Removed AbstractCarrier model implementing the AbstractCarrierInterface. - Changed references from AbstractCarrierInterface to CarrierInterface in de Config model.
