+++ app/code/Magento/Shipping/Model/Carrier/AbstractCarrier.php
@@ -17,7 +17,7 @@ use Magento\Shipping\Model\Shipment\Request;
  * @api
  * @since 100.0.2
  */
-abstract class AbstractCarrier extends \Magento\Framework\DataObject implements AbstractCarrierInterface
+abstract class AbstractCarrier extends \Magento\Framework\DataObject
 {
     const DEBUG_KEYS_MASK = '****';
